+++ CHANGELOG.md
@@ -9,6 +9,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 ### Added
 
 - Added convenience methods for checking the variant of `SeqSrc`
+- Implemented `to_reverse` and `make_reverse` for `MaybeAligned<Alignment<_>>`
 
 ### Fixed